requirements.txt
pyvisa
pyvisa-py
plotly
dash
dash-bootstrap-components
matplotlib
pandas
pyserial
seaborn
fpdf